requirements.txt
groq
flask
python-dotenv
flask-cors
pymongo
werkzeug